write unit tests and refactor code
